Update linter and build process
Updated the package.json to use a higher version of tslint and typescript to allow for some newer things.  Also replaced the tslint file with the default file for now and fixed a couple of little things for that in a couple of the ts files.  Added in a small script to the package.json to make sure that the code compiles and passes the linter.